feat : InterviewOption subcomponent
