+++ src/components/Main/InterviewOptionsPanel/Dropdown/Svg/index.tsx
