+++ src/components/Main/index.tsx
@@ -1,19 +1,14 @@
-import InterviewOptions from './InterviewOptionsPanel';
-import ResponseSection from './ModelResponsePanel';
+import InterviewOptionsPanel from './InterviewOptionsPanel';
+import ModelResponsePanel from './ModelResponsePanel';
 import UserRequestPanel from './UserRequestPanel';
-import { INTERVIEW_OPTIONS } from './constants';
 
 const Main = () => {
   return (
     <main className="flex flex-col items-center gap-12">
-      <div className="flex">
-        {INTERVIEW_OPTIONS.map(option => (
-          <InterviewOptions key={option.id} option={option} />
-        ))}
-      </div>
+      <InterviewOptionsPanel />
       <div className="flex flex-col gap-20">
         <UserRequestPanel />
-        <ResponseSection />
+        <ModelResponsePanel />
       </div>
     </main>
   );
